rsquareclass-master/requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
Pillow==7.0.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
